pyproject.toml
# ──────────────────────────────────────
# 📋 pyproject.toml
# ──────────────────────────────────────
# Python 프로젝트의 "설정 파일"입니다.
# Spring의 pom.xml, Node의 package.json과 같은 역할이에요.
#
# 이 파일 하나로:
#   - 프로젝트 이름, 버전, 설명
#   - 필요한 라이브러리 (dependencies)
#   - 빌드 도구 설정
#   - 개발 도구 설정 (린터, 포매터, 테스트)
# 을 모두 관리합니다.

# ── 빌드 시스템 설정 ──
# "이 프로젝트를 빌드할 때 setuptools를 사용해주세요"
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

# ── 프로젝트 기본 정보 ──
[project]
name = "python-ai-agent"
version = "0.1.0"
description = "파이썬 초보자를 위한 간단한 AI 에이전트"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"

# 필요한 라이브러리 (pip install 할 것들)
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

# 개발할 때만 필요한 라이브러리 (테스트, 린터 등)
[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

# ── 콘솔 명령어 등록 ──
# pip install 후 터미널에서 "ai-agent" 명령어로 실행 가능
[project.scripts]
ai-agent = "ai_agent.main:main"
ai-agent-serve = "ai_agent.api:serve"

# ── Ruff 설정 (코드 린터 & 포매터) ──
# Python의 대표적인 코드 품질 도구
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]  # 에러, 경고, import 정렬

# ── Pytest 설정 (테스트 도구) ──
[tool.pytest.ini_options]
testpaths = ["tests"]
